+++ epsilon-runtime-execution/src/main/java/hu/blackbelt/epsilon/runtime/execution/contexts/EgxExecutionContext.java
@@ -4,7 +4,7 @@ import hu.blackbelt.epsilon.runtime.execution.exceptions.ScriptExecutionExceptio
 import lombok.Builder;
 import lombok.NoArgsConstructor;
 import org.eclipse.epsilon.egl.EgxModule;
-import org.eclipse.epsilon.eol.IEolExecutableModule;
+import org.eclipse.epsilon.eol.IEolModule;
 
 import java.util.List;
 import java.util.Map;
@@ -18,7 +18,7 @@ public class EgxExecutionContext extends EglExecutionContext {
     }
 
     @Override
-    public IEolExecutableModule getModule(Map<Object, Object> context) throws ScriptExecutionException {
+    public IEolModule getModule(Map<Object, Object> context) throws ScriptExecutionException {
         EgxModule module = new EgxModule(getTemplateFactory(context));
         return module;
     }