+++ epsilon-runtime-execution/src/main/java/hu/blackbelt/epsilon/runtime/execution/contexts/EtlExecutionContext.java
@@ -2,8 +2,10 @@ package hu.blackbelt.epsilon.runtime.execution.contexts;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.eclipse.epsilon.eol.IEolExecutableModule;
+import lombok.NonNull;
+import org.eclipse.epsilon.eol.IEolModule;
 import org.eclipse.epsilon.etl.EtlModule;
 
 import java.util.List;
@@ -13,17 +15,32 @@ import java.util.Map;
 @AllArgsConstructor
 public class EtlExecutionContext extends EolExecutionContext {
 
+    @Getter
+    @NonNull
+    private String exportTransformationTrace;
+
     @Builder.Default
     EtlModule etlModule = new EtlModule();
 
     @Builder(builderMethodName = "etlExecutionContextBuilder")
-    public EtlExecutionContext(String source, List<ProgramParameter> parameters) {
+    public EtlExecutionContext(String source, List<ProgramParameter> parameters, String exportTransformationTrace) {
         super(source, parameters);
+        this.exportTransformationTrace = exportTransformationTrace;
     }
 
     @Override
-    public IEolExecutableModule getModule(Map<Object, Object> context) {
+    public IEolModule getModule(Map<Object, Object> context) {
         return etlModule;
     };
+
+    @Override
+    public void post(Map<Object, Object> context) {
+        if (exportTransformationTrace != null) {
+            context.put(
+                    exportTransformationTrace,
+                    etlModule.getContext().getTransformationTrace());
+        }
+    }
+
 
 }